fix(55): cast Decimal p-values to float for BH gate + bulk promote 76 features
PostgreSQL NUMERIC columns return decimal.Decimal objects which scipy's
false_discovery_control rejects. Add .astype(float) before BH correction
in both promoter.py and runner.py. Bulk-promoted 76/100 features to
dim_feature_registry (24 rejected: delta features + constant signals).

+++ src/ta_lab2/experiments/promoter.py
@@ -179,7 +179,9 @@ class FeaturePromoter:
             )
 
         # Apply BH correction to the valid subset
-        adjusted = false_discovery_control(valid_pvals_series.values, method="bh")
+        adjusted = false_discovery_control(
+            valid_pvals_series.values.astype(float), method="bh"
+        )
         df.loc[valid_mask, "ic_p_value_bh"] = adjusted
 
         # Count passing combos
